lass/requirements.txt
lass
Django>=1.4.2
django-blog-zinnia>=0.11.2
django-compressor>=1.2
django-debug-toolbar>=0.9.4
django-endless-pagination==1.2.7
django-flatblocks>=0.6.0
django-less>=0.4
django-mptt>=0.5.4
django-sitetree>=0.9
django-tagging>=0.3.1
django-timedeltafield>=0.64
template_utils>=0.4
xapian-haystack>=1.1.4
